requirements.txt
easyocr
matplotlib
opencv-python-headless

pytesseract
pillow